# api/src/openapi.yaml
openapi: 3.0.2
info:
  version: 0.1.0
  title: NFT Gallery API
  description: API serving the NFT Gallery Platform
...
security:
  - bearerAuth: []
servers:
  - url: '{protocol}://{host}:{port}'
    description: 'Local Mocks (e.g. http://localhost:4010) and Local Dev Server (e.g. http://localhost:3000/v1)'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
        description: Protocol of the URL.
      host:
        enum:
          - api-dev.domi.house
          - api.domi.house
          - localhost
        default: localhost
        description: Host of the URL
      port:
        enum:
          - ''
          - '4010'
          - '3000'
        default: '3000'
        description: Port of the URL
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication token
paths:
...
  /v1/certification:
    post:
      summary: Create Certification
      operationId: createCertification
      x-eov-operation-handler: CertificationsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Certification
      requestBody:
        description: information that is needed for Certification
        required: true
        content:
          application/json:
            schema:
              title: Certification
              type: object
              description: 'Nft Gallery Certification, i.e. an NFT'
              x-autogen:
                openapi: true
                ownership: true
                mongo: true
                security:
                  create:
                    - admin
                    - consumer
                  update:
                    - admin
                    - consumer
                  delete:
                    - admin
                    - consumer
                  retrieve:
                    - admin
                    - consumer
                  retrieves:
                    - admin
                    - consumer
              properties:
                ownerId:
                  type: string
                owner:
                  type: string
                collectionId:
                  type: string
                yearCreated:
                  type: string
                  example: 2021-07-07T10:31:21.072Z
                description:
                  type: string
                  example: New Description
                title:
                  type: string
                  example: New Title
                price:
                  type: string
                  example: 1.5
                likes:
                  type: string
                  example: 200
                  minimum: 0
                unencryptedFiles:
                  type: array
                  items:
                    type: string
                    format: binary
                encryptedFiles:
                  type: array
                  items:
                    type: object
                    properties:
                      path:
                        type: string
                      content:
                        type: string
                royalty:
                  type: string
                  example: 5
                properties:
                  type: string
                saleType:
                  type: string
                  enum:
                    - PutOnSale
                    - InstantSalePrice
                    - UnlockOnePurchased
              required:
                - owner
                - title
                - description
                - saleType
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
  /v1/certifications:
    parameters:
      - name: limit
        in: query
        description: How many items to return at one time (max 100)
        required: false
        schema:
          type: integer
          format: int32
      - name: skip
        in: query
        description: How many items should be skipped before returnig the required amount
        required: false
        schema:
          type: integer
          format: int32
    get:
      summary: Get all Certifications
      operationId: retrieveCertifications
      x-eov-operation-handler: CertificationsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Certification
      parameters:
        - name: ownerId
          in: query
          required: false
          description: The Fileter ownerId
          schema:
            type: string
            example: null
        - name: owner
          in: query
          required: false
          description: The Fileter owner
          schema:
            type: string
            example: null
        - name: collectionId
          in: query
          required: false
          description: The Fileter collectionId
          schema:
            type: string
            example: null
        - name: yearCreated
          in: query
          required: false
          description: The Fileter yearCreated
          schema:
            type: string
            example: 'Wed Jul 07 2021 12:31:21 GMT+0200 (Central European Summer Time)'
        - name: description
          in: query
          required: false
          description: The Fileter description
          schema:
            type: string
            example: New Description
        - name: title
          in: query
          required: false
          description: The Fileter title
          schema:
            type: string
            example: New Title
        - name: price
          in: query
          required: false
          description: The Fileter price
          schema:
            type: string
            example: 1.5
        - name: likes
          in: query
          required: false
          description: The Fileter likes
          schema:
            type: string
            example: 200
        - name: unencryptedFiles
          in: query
          required: false
          description: The Fileter unencryptedFiles
          schema:
            type: array
            example: null
        - name: encryptedFiles
          in: query
          required: false
          description: The Fileter encryptedFiles
          schema:
            type: array
            example: null
        - name: royalty
          in: query
          required: false
          description: The Fileter royalty
          schema:
            type: string
            example: 5
        - name: properties
          in: query
          required: false
          description: The Fileter properties
          schema:
            type: string
            example: null
        - name: saleType
          in: query
          required: false
          description: The Fileter saleType
          schema:
            type: string
            default: null
            enum:
              - PutOnSale
              - InstantSalePrice
              - UnlockOnePurchased
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
  '/v1/certifications/{certificationId}':
    parameters:
      - name: certificationId
        in: path
        required: true
        description: The Id of the Certification
        schema:
          description: Id of the entity
          type: string
          format: not-empty-string
          example: 60758e80d59c357138d01af6
    get:
      summary: Get Certification with Id
      operationId: retrieveCertification
      x-eov-operation-handler: CertificationsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Certification
      parameters:
        - name: certificationId
          in: path
          required: true
          description: The Id of the Certification
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
    put:
      summary: Update Certification with Id
      operationId: updateCertification
      x-eov-operation-handler: CertificationsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Certification
      parameters:
        - name: certificationId
          in: path
          required: true
          description: The Id of the Certification
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      requestBody:
        description: information that is needed for Certification
        required: true
        content:
          application/json:
            schema:
              title: Update Certification with Id
              type: object
              description: An array of fields that we want to change name and value
              properties:
                ownerId:
                  type: string
                owner:
                  type: string
                collectionId:
                  type: string
                yearCreated:
                  type: string
                  example: 'Wed Jul 07 2021 12:31:21 GMT+0200 (Central European Summer Time)'
                description:
                  type: string
                  example: New Description
                title:
                  type: string
                  example: New Title
                price:
                  type: string
                  example: 1.5
                likes:
                  type: string
                  example: 200
                unencryptedFiles:
                  type: array
                encryptedFiles:
                  type: array
                royalty:
                  type: string
                  example: 5
                properties:
                  type: string
                saleType:
                  type: string
                  enum:
                    - PutOnSale
                    - InstantSalePrice
                    - UnlockOnePurchased
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
    delete:
      summary: Delete Certification with Id
      operationId: deleteCertification
      x-eov-operation-handler: CertificationsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Certification
      parameters:
        - name: certificationId
          in: path
          required: true
          description: The Id of the Certification
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common